refactor: move wrapController out of app.ts
